requirements.txt
python-telegram-bot==21.3
google-generativeai==0.7.0